Add style to ButtonPanel
